Conflicts:
	src/main/java/com/devicehive/controller/UserController.java
	src/main/java/com/devicehive/dao/NetworkDAO.java
	src/main/java/com/devicehive/dao/UserDAO.java
	src/main/java/com/devicehive/model/Equipment.java
	src/main/java/com/devicehive/service/NetworkService.java

+++ src/main/java/com/devicehive/service/DeviceService.java
@@ -1,11 +1,9 @@
 package com.devicehive.service;
 
-import com.devicehive.configuration.Constants;
 import com.devicehive.dao.*;
 import com.devicehive.exceptions.HiveException;
 import com.devicehive.model.*;
-import com.devicehive.service.interceptors.ValidationInterceptor;
-import com.devicehive.websockets.json.GsonFactory;
+import com.devicehive.json.GsonFactory;
 import com.devicehive.websockets.messagebus.global.MessagePublisher;
 import com.devicehive.websockets.messagebus.local.LocalMessageBus;
 import com.google.gson.Gson;
@@ -16,20 +14,15 @@ import org.slf4j.LoggerFactory;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
-import javax.interceptor.Interceptors;
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import javax.websocket.Session;
+import java.sql.Timestamp;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
-@Interceptors(ValidationInterceptor.class)
 @Stateless
 public class DeviceService {
     private static final Logger logger = LoggerFactory.getLogger(DeviceService.class);
-    @PersistenceContext(unitName = Constants.PERSISTENCE_UNIT)
-    private EntityManager em;
     @Inject
     private DeviceCommandDAO deviceCommandDAO;
     @Inject
@@ -50,15 +43,17 @@ public class DeviceService {
     private DeviceEquipmentDAO deviceEquipmentDAO;
 
     public void deviceSave(Device device, Set<Equipment> equipmentSet) {
+        if (device.getNetwork() != null) {
             device.setNetwork(networkService.createOrVeriryNetwork(device.getNetwork()));
-            device.setDeviceClass(createOrUpdateDeviceClass(device.getDeviceClass(), equipmentSet));
-            createOrUpdateDevice(device);
+        }
+        device.setDeviceClass(createOrUpdateDeviceClass(device.getDeviceClass(), equipmentSet));
+        createOrUpdateDevice(device);
     }
 
     public void submitDeviceCommand(DeviceCommand command, Device device, User user, Session userWebsocketSession) {
         command.setDevice(device);
         command.setUser(user);
-        command.setTimestamp(new Date());
+        command.setTimestamp(new Timestamp(System.currentTimeMillis()));
         deviceCommandDAO.saveCommand(command);
         if (userWebsocketSession != null) {
             localMessageBus.subscribeForCommandUpdates(command.getId(), userWebsocketSession);
@@ -97,7 +92,8 @@ public class DeviceService {
                                                            DeviceEquipment deviceEquipment) {
         if (deviceEquipment != null) {
             if (deviceEquipmentDAO.update(deviceEquipment) == 0) {
-                logger.warn("No equipments to update found");
+                deviceEquipment.setTimestamp(new Timestamp(System.currentTimeMillis()));
+                deviceEquipmentDAO.saveDeviceEquipment(deviceEquipment);
             }
         }
         notification.setDevice(device);
@@ -119,7 +115,7 @@ public class DeviceService {
     public DeviceClass createOrUpdateDeviceClass(DeviceClass deviceClass, Set<Equipment> newEquipmentSet) {
         DeviceClass stored;
         if (deviceClass.getId() != null) {
-            stored = em.find(DeviceClass.class, deviceClass.getId());
+            stored = deviceClassDAO.getDeviceClass(deviceClass.getId());
         } else {
             stored = deviceClassDAO.getDeviceClassByNameAndVersion(deviceClass.getName(),
                     deviceClass.getVersion());
@@ -135,7 +131,7 @@ public class DeviceService {
             return stored;
         } else {
             //create
-            em.persist(deviceClass);
+            deviceClassDAO.saveDeviceClass(deviceClass);
             updateEquipment(newEquipmentSet, deviceClass);
             return deviceClass;
         }
@@ -144,28 +140,19 @@ public class DeviceService {
 
     public void updateEquipment(Set<Equipment> newEquipmentSet, DeviceClass deviceClass) {
         List<Equipment> existingEquipments = equipmentDAO.getByDeviceClass(deviceClass);
-        if (!newEquipmentSet.isEmpty()) {
-            if (!existingEquipments.isEmpty()) {
-                equipmentDAO.removeEquipment(existingEquipments);
-            } else {
-                for (Equipment equipment : newEquipmentSet) {
-                    DeviceEquipment deviceEquipment = new DeviceEquipment();
-                    deviceEquipment.setTimestamp(new Date(0));
-                    deviceEquipment.setCode(equipment.getCode());
-                    em.persist(deviceEquipment);
-                }
-            }
+        if (!newEquipmentSet.isEmpty() && !existingEquipments.isEmpty()) {
+            equipmentDAO.removeEquipment(existingEquipments);
         }
         for (Equipment equipment : newEquipmentSet) {
             equipment.setDeviceClass(deviceClass);
-            em.persist(equipment);
+            equipmentDAO.saveEquipment(equipment);
         }
     }
 
     public void createOrUpdateDevice(Device device) {
-        Device existingDevice = deviceDAO.findByUUIDForWrite(device.getGuid());
+        Device existingDevice = deviceDAO.findByUUID(device.getGuid());
         if (existingDevice == null) {
-            em.persist(device);
+            deviceDAO.saveDevice(device);
         } else {
             existingDevice.setDeviceClass(device.getDeviceClass());
             existingDevice.setStatus(device.getStatus());
